Update Ephemeral Storage Fee
Withdraw. We leverage the fact that AIP-57 imposes a upper bound to the disk bandwidth in practice (hence a upper bound to the total disk consumption for a prune window in terms of wall time) and remove the ephemeral bytes fee at all. With the simplified model, an AIP describing purely the refundable bytes will be drafted in place of this.